requirements.txt
# Core dependencies
SpeechRecognition>=3.8.1
pyttsx3>=2.90
openai>=0.27.0
PyAudio>=0.2.11
python-dotenv>=0.19.0
requests>=2.28.0

# Additional utilities
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

# Web functionality
urllib3>=1.26.0

# System utilities
psutil>=5.9.0

# New API integrations
beautifulsoup4>=4.11.0
lxml>=4.9.0

# GUI enhancements
pillow>=9.0.0